requirements.txt
absl-py==1.4.0
altgraph==0.17.3
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1688651106312/work/dist
appscript==1.2.5
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1671589829431/work
aspose-words==24.10.0
astunparse==1.6.3
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764961872/work
cachetools==5.3.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179414629/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1690553059027/work
conda-content-trust @ file:///home/conda/feedstock_root/build_artifacts/conda-content-trust_1621370699668/work
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1669907009957/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1669733752472/work
contourpy==1.0.7
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1677821361175/work
cycler==0.11.0
DateTime==5.2
Deprecated==1.2.14
distro==1.8.0
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1683130392307/work
django-cors-headers==4.0.0
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1663854154474/work
docx2pdf==0.1.8
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/fastapi_1692014902720/work
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.4
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httptools==0.6.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
jax==0.4.10
Jinja2==3.1.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1632759296524/work
jsonpointer==2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==5.3.0
macholib==1.16.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
netaddr==0.9.0
networkx==3.1
ngrok==0.9.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1691056334969/work
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.1
Pillow==9.5.0
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1667232663820/work
protobuf==4.23.1
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes @ file:///home/conda/feedstock_root/build_artifacts/pybboxes_1672656586717/work
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1666836580084/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic @ file:///Users/runner/miniforge3/conda-bld/pydantic_1690476298755/work
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1672659226110/work
pypandoc==1.14
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1666772684107/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1678273079927/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1670412871362/work
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.23.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1662051266223/work
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1681817562700/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette-recipe_1684245096404/work
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-macos==2.12.0
termcolor==2.2.0
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==2.0.1
torchvision==0.15.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1677948868469/work
typing_extensions==4.12.2
tzdata==2023.3
ultralytics==8.0.105
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.14.1
yookassa==2.4.0
zope.interface==6.0
zstandard==0.19.0
